extract movable-point-editor
